Prevent null user reference + remove unused function
When a user is entered in the same Room and MMORoom as another player and they disconnect, `fnUserLost` forcefully removes the user from the GlobalUserManager. The MMORoom receives a `fnSetUserPosition` request after, breaking the connection when trying to evaluate a user that no longer exists.

`fnUserLost` now respects the user duplicate count like other functions, which removes the need for the separate `removeUserReference` function.